Add files via upload
HOG focuses on the structure of the object. It extracts the information of the edges magnitude as well as the orientation of the edges.

It uses a detection window of 64x128 pixels, so the image is first converted into (64, 128) shape.

The image is then further divided into small parts, and then the gradient and orientation of each part is calculated. It is divided into 8x16 cells into blocks with 50% overlap, so there are going to be 7x15 = 105 blocks in total, and each block consists of 2x2 cells with 8x8 pixels.

We take the 64 gradient vectors of each block (8x8 pixel cell) and put them into a 9-bin histogram.